+++ R/list.R
@@ -5,7 +5,7 @@
 #' @param values An optional vector or named list specifying the values.
 #' @param unique A flag indicating whether the values must be unique.
 #' @param sorted A flag indicating whether the list must be sorted.
-#' @param named A flag indicating whether the list must be named or unnamed or a regular expression that must match all the names or NA if it doesn't matter if the list is named.
+#' @param named A flag indicating whether the list must be named or unnamed or a regular expression that must match all the names or count or count range of the number of characters in the names or NA if it doesn't matter if the list is named.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
@@ -22,15 +22,19 @@ check_list <- function(x,
   
   if (!is.list(x)) error(x_name, " must be a list")
 
-  if(!(is_flag(named) || is_string(named) || is_NA(named))) 
-    error("named must be a flag a string or NA")
+  if(!(is_flag(named) || is_string(named) || is_NA(named) || is_count(named) || is_count_range(named))) 
+    error("named must be a flag, string, count, count range or NA")
   
   regex <- ".*"
+  nchar <- c(0L, .Machine$integer.max)
   if(is_string(named)) {
     regex <- named
     named <- TRUE
+  } else if(is_count(named) || is_count_range(named)) {
+    nchar <- named
+    named <- TRUE
   }
-  
+    
   check_flag_internal(error)
   
   if(!missing(values)) {
@@ -61,7 +65,7 @@ check_list <- function(x,
   if(sorted) check_sorted(x, x_name = x_name, error = error)
   
   if(is_flag(named) && named) {
-    check_named(x, regex = regex, x_name = x_name, error = error)
+    check_named(x, nchar = nchar, regex = regex, x_name = x_name, error = error)
   } else if(is_flag(named) && !named)
     check_unnamed(x, x_name = x_name, error = error)