+++ R/named.R
@@ -1,6 +1,7 @@
 #' Check Named
 #'
 #' @param x The object to check.
+#' @param nchar A count or count range of the number of characters.
 #' @param regex A string of the regular expression that must match all names.
 #' @param unique A flag indicating whether the names must be unique.
 #' @param x_name A string of the name of the object.
@@ -13,17 +14,21 @@
 #' x <- 1
 #' names(x) <- "y"
 #' check_named(x, error = FALSE)
-check_named <- function(x, regex = ".*", unique = FALSE, 
-                         x_name = substitute(x),
-                         error = TRUE) {
+check_named <- function(x, nchar = c(0L, .Machine$integer.max), 
+                        regex = ".*", unique = FALSE, 
+                        x_name = substitute(x),
+                        error = TRUE) {
   x_name <- deparse_x_name(x_name)
-
+  
   check_flag(unique)
   check_flag_internal(error)
-
+  
   if(is.null(names(x))) {
     on_fail(x_name, " must be named", error = error)
-  } else check_regex(names(x), regex = regex, x_name = paste("names of", x_name), error = error)
+  } else {
+    check_nchar(names(x), nchar = nchar, x_name = paste("names of", x_name), error = error)
+    check_regex(names(x), regex = regex, x_name = paste("names of", x_name), error = error)
+  }
   if(unique)
     check_unique(names(x), x_name = paste("names of", x_name), error = error)
   invisible(x)