+++ NEWS.md
@@ -1,6 +1,6 @@
 - added `check_nchar()` function
-- `check_vector()` and `check_list()` now allow `named` argument to be a regular expression
-- added `regex = ".*"` argument to `check_named()`
+- `check_vector()` and `check_list()` now allow `named` argument to be a regular expression or count range
+- added `nchar = c(0L, .Machine$max.integer)` and `regex = ".*"` arguments to `check_named()`
 - added `check_regex()` function
 - added `all_y = TRUE` argument to `check_join()` to check all rows in y in join
 - changed `check_join()` error message to